refactor" reshuffling components in dedicated _components folder
